version 20200609
version 20200609, element of type "FUNCTIONS" added  (mind the S at the end, it is not the same as "FUNCTION"). It allows to compile multiple functionS in one FCMP call. Description/instruction will be added soon.
The code of each new function in 'xxx_functions' subfolder should be stored in a single file but without "proc fcmp" header and "run" footer just plain function/subroutine definition.