macroArray v.0.4
New parameter added to the %do_over macro.

The `which=` is a _SPACE_ separated list of indexes which should be used, possible special characters are `H` and `L` which means high and low bound of an array, list could be set with colons(:) in form of start:end:by (no spaces between!), if by is omitted the default is 1. If possible use `1:5` rather `1 2 3 4 5` since the firs works faster.