macroArray package, version 0.6
macroArray package, version 0.6:

New feature of the `%array()` macro. If the `macarray` parameter of the `%array()` macro is set to `M` then for a given array name the macro symbols table is scanned for macrovariables with prefix like the array name and numeric suffixes. Then the minimum and the maximum index is determined and all not existing global macrovariables are created and a macro is generated in the same way as it is generated for the `Y` value.

Documentation updated.