macroArray, version 0.7:
macroArray, version 0.7:

The `%mcHashTable()` macro was added in the package.
It is designed to facilitate the idea of a "macro hash table"
concept, i.e. *a list of macrovariables with common prefix
and suffixes generated as a hash digest* which allows
to use values other than integers as indexes.
The `%mcHashTable()` macro allows to generate other macros
which behaves like hash tables or dictionaries.